Split backup.py into python.py & BackupServices class
Idea is one script in the root directory which just takes care of all
the commandline argument stuff.  This then initialises the
BackupServices class which orchestrates the backup.

Also, now generate the services registry in a class method on the class.
This makes sure that --list and the actual backup action have the same
idea of which services are available