Add User configurable timestamps for backups
This is accomplished via the 'timestamp' configuration parameters, which
defaults to '%Y-%m-%d' as before.  The time used in the stamp can be
overridden, but defaults to the current time.

To get this to work, I had to disable
[interpolation](https://docs.python.org/3.4/library/configparser.html#interpolation-of-values)
of ConfigParser files, but this seems like a feature we're unlikely to
use anyway?